requirements.txt
aniso8601==9.0.1
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
cryptography==3.4.7
Flask==2.0.1
Flask-Limiter==1.4
flask-restx==0.5.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
jwt==1.2.0
limits==1.5.1
lxml==4.6.3
MarkupSafe==2.0.1
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pyrsistent==0.18.0
python-cas==1.5.0
python-ldap==3.3.1
pytz==2021.1
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.22
urllib3==1.26.6
Werkzeug==2.0.1
